refactor(functions): separate firestore tokens utils
