+++ packages/functions/src/functions/firestore/onUpdateUser.ts
@@ -3,7 +3,9 @@ import * as functions from 'firebase-functions';
 import * as Twitter from 'twitter';
 import * as _ from 'lodash';
 import { env } from '../../utils/env';
-import { checkInvalidToken, setTokenInvalid, getToken, setRecord, existsRecords } from '../../utils/firestore';
+import { checkInvalidToken, setRecord, existsRecords } from '../../utils/firestore';
+import { getToken } from '../../utils/firestore/tokens/getToken';
+import { setTokenInvalid } from '../../utils/firestore/tokens/setTokenInvalid';
 import { getTwUsers } from '../../utils/firestore/twUsers/getTwUsers';
 import { setTwUsers } from '../../utils/firestore/twUsers/setTwUsers';
 import { getUsersLookup } from '../../utils/twitter';
